migrate over to a model
