--- simple_note_taker/database/notes.py
@@ -1,7 +1,0 @@
-# from simple_note_taker.database.core import tiny_db
-#
-#
-# class NotesDB:
-#     db = tiny_db.table("notes")
-#
-#     def searchNotes(self) -> :
